Update style, use safer casting
This commit updates the the algorithm to use
spaces instead of tabs. Also other minor formatting
changes were made for ease of readability and for
variable scoping. Finally, the inverse square root
function was updated to use a safer cast from float
to int32 than what was previously being used.